修改一些细枝末节+将Ch8的VkBase.h和VkBase+.h更新到更接近我个人使用的版本

+++ solution/EasyVulkan_Ch6/VKBase+.h
@@ -93,7 +93,7 @@ namespace vulkan {
 			outStream << std::format("[ FormatInfo ] ERROR\nThis function only supports definite formats provided by VK_VERSION_1_0.\n"),
 			abort();
 #endif
-		return formatInfos_v1_0[uint32_t(format)];
+		return formatInfos_v1_0[format];
 	}
 	constexpr VkFormat Corresponding16BitFloatFormat(VkFormat format_32BitFloat) {
 		switch (format_32BitFloat) {
@@ -280,7 +280,7 @@ namespace vulkan {
 			};
 			VkResult result = vkWaitSemaphores(graphicsBase::Base().Device(), &waitInfo, UINT64_MAX);
 			if (result)
-				outStream << std::format("[ timelineSemaphore ] ERROR\nFailed to wait for semaphores!\nError code: {}\n", int32_t(result));
+				outStream <<std::format("[ timelineSemaphore ] ERROR\nFailed to wait for semaphores!\nError code: {}\n", int32_t(result));
 			return result;
 		}
 	};
